adding categories and subcategories, deleting payments, scheduled service

+++ src/main/java/org/knop/budgetKeeper/repository/PaymentRepository.java
@@ -1,15 +1,15 @@
 package org.knop.budgetKeeper.repository;
 
-import org.knop.budgetKeeper.dto.PaymentDto;
 import org.knop.budgetKeeper.models.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.sql.Date;
 import java.util.List;
 
 @Repository
-public interface PaymentRepository extends JpaRepository<Payment, Integer> {
+public interface PaymentRepository extends JpaRepository<Payment, BigInteger> {
     List<Payment> findAllByUserId(Integer userId);
 
     List<Payment> findAllByUserIdAndIncomeLabelAndDateBetween(Integer userId,
